Refactor prototype single-file picking functionality, feature #45
Moved most of BrowserActivity code into AbstractBrowserActivity,
which is then extended by BrowserActivity and the new PickerActivity.
BrowserFragment still uses a boolean and conditionals; however, this
commit clearly demonstrates the refactored approach.

Base tests pass, file picking works.

Need to complete refactor on BrowserFragment, debug, and add tests
for picking.
