+++ explorer/src/main/java/com/dnielfe/manager/adapters/BrowserTabsAdapter.java
@@ -2,7 +2,6 @@ package com.dnielfe.manager.adapters;
 
 import android.app.Fragment;
 import android.app.FragmentManager;
-import android.os.Bundle;
 import android.support.v13.app.FragmentStatePagerAdapter;
 import android.view.ViewGroup;
 
@@ -15,11 +14,8 @@ public class BrowserTabsAdapter extends FragmentStatePagerAdapter {
 
     private static Fragment mCurrentFragment;
 
-    private boolean mIsGetContent;
-
-    public BrowserTabsAdapter(FragmentManager fm, boolean isGetContent) {
+    public BrowserTabsAdapter(FragmentManager fm) {
         super(fm);
-        mIsGetContent = isGetContent;
     }
 
     public static BrowserFragment getCurrentBrowserFragment() {
@@ -28,11 +24,7 @@ public class BrowserTabsAdapter extends FragmentStatePagerAdapter {
 
     @Override
     public Fragment getItem(int pos) {
-        final BrowserFragment bf = new BrowserFragment();
-        final Bundle args = new Bundle();
-        args.putBoolean(BrowserFragment.KEY_IS_GET_CONTENT, mIsGetContent);
-        bf.setArguments(args);
-        return bf;
+        return new BrowserFragment();
     }
 
     @Override
