+++ explorer/src/main/res/layout/directory_buttons.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<HorizontalScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/scroll_text"
+    android:layout_width="match_parent"
+    android:layout_height="@dimen/nav_height"
+    android:fillViewport="true"
+    android:background="?colorPrimary"
+    android:scrollbars="none">
+
+    <LinearLayout
+        android:id="@+id/directory_buttons"
+        android:layout_width="wrap_content"
+        android:layout_height="match_parent"
+        android:gravity="center_vertical"
+        android:longClickable="true"
+        android:orientation="horizontal" />
+</HorizontalScrollView>
